backend/requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-telegram-bot==21.9
supabase==2.10.0
python-dotenv==1.0.1
pytz==2024.2
python-dateutil==2.9.0
pydantic==2.10.4
httpx==0.28.1
aiosmtplib==3.0.2
python-multipart==0.0.19
apscheduler==3.10.4
